OGM-557 Move unit test to core. Use FEST assertThat. Rename snapshot classes.
