+++ core/src/test/java/org/hibernate/ogm/backendtck/queries/AnEmbeddable.java
@@ -4,13 +4,12 @@
  * License: GNU Lesser General Public License (LGPL), version 2.1 or later
  * See the lgpl.txt file in the root directory or <http://www.gnu.org/licenses/lgpl-2.1.html>.
  */
-package org.hibernate.ogm.datastore.mongodb.test.query;
+package org.hibernate.ogm.backendtck.queries;
 
 import javax.persistence.Embeddable;
 
 @Embeddable
 public class AnEmbeddable {
-
 	String embeddedString;
 
 	public AnEmbeddable() {
@@ -18,6 +17,10 @@ public class AnEmbeddable {
 
 	public AnEmbeddable(String embeddedString) {
 		this.embeddedString = embeddedString;
+	}
+
+	public String getEmbeddedString() {
+		return embeddedString;
 	}
 
 }
