+++ core/src/test/java/org/hibernate/ogm/backendtck/queries/WithEmbedded.java
@@ -4,15 +4,19 @@
  * License: GNU Lesser General Public License (LGPL), version 2.1 or later
  * See the lgpl.txt file in the root directory or <http://www.gnu.org/licenses/lgpl-2.1.html>.
  */
-package org.hibernate.ogm.datastore.mongodb.test.query;
+package org.hibernate.ogm.backendtck.queries;
 
 import javax.persistence.Embedded;
 import javax.persistence.Entity;
 import javax.persistence.Id;
+import org.hibernate.search.annotations.Field;
+import org.hibernate.search.annotations.Indexed;
 
+@Indexed
 @Entity
 public class WithEmbedded {
 
+	@Field
 	@Id
 	Long id;
 
@@ -25,6 +29,10 @@ public class WithEmbedded {
 	public WithEmbedded(Long id, AnEmbeddable anEmbeddable) {
 		this.id = id;
 		this.anEmbeddable = anEmbeddable;
+	}
+
+	public AnEmbeddable getAnEmbeddable() {
+		return anEmbeddable;
 	}
 
 }
